+++ imaginerium/pages/dashboard.js
@@ -1,0 +1,71 @@
+import Head from "next/head";
+import Image from "next/image";
+import { Inter } from "@next/font/google";
+import styles from "../styles/Home.module.css";
+import Link from "next/link";
+import Navigation from "../Components/Navigation";
+import ButtonLink from "../Components/ButtonLink";
+import { VscPersonAdd, VscOrganization
+} from "react-icons/vsc";
+
+
+const inter = Inter({ subsets: ["latin"] });
+
+export default function Home() {
+  return (
+    <>
+      <Head>
+        <title>Imaginerium</title>
+        <meta name="description" content="Generated by create next app" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <div className={styles.homeContainer}>
+      <Navigation />
+      <main className={styles.main}>
+      <div className={styles.titleContainer}>
+        <h1>IMAGINERIUM</h1>
+        <div className={styles.descriptionContainer}>
+          <Image
+            priority
+              src="https://i.ibb.co/ZVj6bcP/spinning-Earth-unscreen.gif"
+            className={styles.earth}
+            height={320}
+            width={ 427}
+            alt="Earth"
+          />
+         
+          <p>
+            Imaginerium is a place where you can create your own characters and
+            share them with the world.
+          </p>
+          
+        </div>
+
+        <div className={styles.buttonsContainer}>
+       
+          <button className={styles.buttonMain}>
+          <Link href="/createCharacter"> 
+          <div>
+            Create Character
+            </div>
+            <VscPersonAdd className={styles.add}/>
+            </Link>
+          </button>
+        
+          
+         <button className={styles.buttonMain}><Link href="/characterList" >
+          <div className={styles.buttonMain}>
+          Character List
+          </div>
+        
+            <VscOrganization className={styles.add}/></Link>
+          </button>
+
+        </div>
+        </div>
+      </main>
+      </div>
+    </>
+  );
+}
