+++ imaginerium/pages/index.js
@@ -1,71 +1,70 @@
-import Head from "next/head";
-import Image from "next/image";
-import { Inter } from "@next/font/google";
-import styles from "../styles/Home.module.css";
-import Link from "next/link";
-import Navigation from "../Components/Navigation";
-import ButtonLink from "../Components/ButtonLink";
-import { VscPersonAdd, VscOrganization
-} from "react-icons/vsc";
+import { signIn } from "next-auth/react";
+import GoogleProvider from "next-auth/providers/google";
+import GithubProvider from "next-auth/providers/github";
+import styles from "../styles/landing.module.css";
 
+export default function landing(){
 
-const inter = Inter({ subsets: ["latin"] });
+return (
 
-export default function Home() {
-  return (
-    <>
-      <Head>
-        <title>Imaginerium</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <div className={styles.homeContainer}>
-      <Navigation />
-      <main className={styles.main}>
-      <div className={styles.titleContainer}>
-        <h1>IMAGINERIUM</h1>
-        <div className={styles.descriptionContainer}>
-          <Image
-            priority
-              src="https://i.ibb.co/ZVj6bcP/spinning-Earth-unscreen.gif"
-            className={styles.earth}
-            height={320}
-            width={ 427}
-            alt="Earth"
-          />
-         
-          <p>
-            Imaginerium is a place where you can create your own characters and
-            share them with the world.
-          </p>
-          
-        </div>
-
-        <div className={styles.buttonsContainer}>
-       
-          <button className={styles.buttonMain}>
-          <Link href="/createCharacter"> 
-          <div>
-            Create Character
-            </div>
-            <VscPersonAdd className={styles.add}/>
-            </Link>
-          </button>
-        
-          
-         <button className={styles.buttonMain}><Link href="/characterList" >
-          <div className={styles.buttonMain}>
-          Character List
+<>
+<div className={styles.body}>
+        <div className={styles.scene}>
+        <div className={styles.upper}>
+          <img onClick={() => signIn(GoogleProvider.id || GithubProvider.id, {
+      callbackUrl: `${window.location.origin}/dashboard`,
+    })} className={styles.earth} src="./earth1.png"/>
+          <img className={styles.moon1copy} src="./moon1.png"/>
+          <img className={styles.moon1} src="./moon1.png"/>
+            <img className={styles.moon2copy} src="./moon3.png"/>
+            <img className={styles.moon2} src="./moon3.png"/>
           </div>
-        
-            <VscOrganization className={styles.add}/></Link>
-          </button>
-
-        </div>
-        </div>
-      </main>
+          
+          <div className={styles.star1}></div>
+          <div className={styles.star2}></div>
+          <div className={styles.star3}></div>
+          <div className={styles.star4}></div>
+          <div className={styles.star5}></div>
+          </div>
       </div>
+      <div >
+      <div className={styles.title}>
+        <div className={styles.hover}>
+          <h1>I</h1>
+        </div>
+        <div className={styles.hover}>
+          <h1>M</h1>
+        </div>
+        <div className={styles.hover}>
+          <h1>A</h1>
+        </div>
+        <div className={styles.hover}>
+          <h1>G</h1>
+        </div>
+        <div className={styles.hover}>
+          <h1>I</h1>
+        </div>
+        <div className={styles.hover}>
+          <h1>N</h1>
+        </div>
+        <div className={styles.hover}>
+          <h1>E</h1>
+        </div>
+        <div className={styles.hover}>
+          <h1>R</h1>
+        </div>
+        <div className={styles.hover}>
+          <h1>I</h1>
+        </div>
+        <div className={styles.hover}>
+          <h1>U</h1>
+        </div>
+        <div className={styles.hover}>
+          <h1>M</h1>
+        </div>
+      </div>
+    </div>
+   
     </>
-  );
+    )
 }
